Refactor tests to follow minitest
