+++ test/fixtures/files/vc-c19-pcr-jwt-payload.json
